Projeto Java/base.core/src/test/java/eapli/base/AGV/domain/BriefDescriptionTest.java
package eapli.base.AGV.domain;

import eapli.base.warehouse.domain.AGVDock;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

class BriefDescriptionTest {

    @Test
    void getBriefDescription() {
        AGVId id = new AGVId("12345678");
        BriefDescription description = new BriefDescription("abcdefg");
        MaxWeightCapacity weight = new MaxWeightCapacity(200.0);
        MaxVolumeCapacity volume = new MaxVolumeCapacity(200);
        Model model = new Model("2.1.1.1");
        Range a = new Range(5.0);
        AGVPosition pos = new AGVPosition("s");
        AGVDock dock = new AGVDock();
        AGVStatus agvStatus = new AGVStatus(AGVStatus.Status.FREE);

        AGV a1 = new AGV(id,description,model,weight,volume,a,pos,dock, agvStatus);
        AGV a2 = new AGV(id,description,model,weight,volume,a,pos,dock, agvStatus);
        Assertions.assertEquals(a1.getBriefDescription(), a2.getBriefDescription());
    }

    @Test
    void setBriefDescription() {
        AGVId id = new AGVId("12345678");
        BriefDescription description = new BriefDescription("abcdefg");
        MaxWeightCapacity weight = new MaxWeightCapacity(200.0);
        MaxVolumeCapacity volume = new MaxVolumeCapacity(200);
        Model model = new Model("2.1.1.1");
        Range a = new Range(5.0);
        AGVPosition pos = new AGVPosition("s");
        AGVDock dock = new AGVDock();
        AGVStatus agvStatus = new AGVStatus(AGVStatus.Status.FREE);

        AGV a1 = new AGV(id,description,model,weight,volume,a,pos,dock, agvStatus);
        description.setBriefDescription("abcdefg");
        AGV a2 = new AGV(id,description,model,weight,volume,a,pos,dock, agvStatus);
        Assertions.assertEquals(a1.getBriefDescription(), a2.getBriefDescription());
    }

    @Test
    void testToString() {
        BriefDescription m2 = new BriefDescription("1234");
        String expected = "BriefDescription{briefDescription='1234'}";
        Assertions.assertEquals( m2.toString(),expected);
    }
    @Test
    void testToString2() {
        BriefDescription m = new BriefDescription();
        BriefDescription m2 = new BriefDescription("1234");
        Assertions.assertNotEquals(m.toString(), m2.toString());
    }
}